pyproject.toml
[project]
name = "multimodal-ai"
version = "0.1.0"
description = "Visual Question Answering with ViLT"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "peft>=0.18.0",
    "pillow>=11.3.0",
    "timm>=1.0.22",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "transformers>=4.57.1",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.1.0",
]

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]
